% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{app_server}
\alias{app_server}
\title{The main server function for this shiny app}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{input object}

\item{output}{output object}
}
\description{
The main server function for this shiny app
}
