% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mod.R
\name{slider_text_server}
\alias{slider_text_server}
\title{The slider mod server}
\usage{
slider_text_server(input, output, session)
}
\arguments{
\item{input}{input class object}

\item{output}{output class object}

\item{session}{}
}
\value{
a reactive procedure/set of instructions
}
\description{
The slider mod server
}
