% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mod.R
\name{slider_text_ui}
\alias{slider_text_ui}
\title{The slider modular ui}
\usage{
slider_text_ui(id)
}
\arguments{
\item{id}{the relevant prefix}
}
\value{
html that builds the slider
}
\description{
The slider modular ui
}
